+++ lib/flexible_scrollable_table_view.dart
@@ -11,19 +11,22 @@ export 'src/arguments/table_row_build_arguments.dart';
 export 'src/constraint/flexible_table_column_width.dart';
 export 'src/constraint/flexible_table_row_height.dart';
 
-export 'src/decoration/flexible_table_decorations.dart';
+export 'src/custom/cache/table_cell_widget_cache_mixin.dart';
 
 // export 'src/custom/dynamic_width/dynamic_width.dart';
 // export 'src/custom/dynamic_width/dynamic_width_column.dart';
 // export 'src/custom/dynamic_width/intrinsic_width_group.dart';
 // export 'src/custom/dynamic_width/size_listenable_widget.dart';
 
-export 'src/custom/cache/table_cell_widget_cache_mixin.dart';
-
 export 'src/custom/flexible_table_controller.dart';
+
+export 'src/decoration/flexible_table_decorations.dart';
 
 export 'src/layout_builder/lazy_layout_builder.dart';
 export 'src/layout_builder/lazy_sliver_layout_builder.dart';
+
+export 'src/row/table_header_row_build_mixin.dart';
+export 'src/row/table_info_row_build_mixin.dart';
 
 export 'src/scrollable/scroll_behavior.dart';
 export 'src/scrollable/synchronized_scroll_mixin.dart';
@@ -40,10 +43,6 @@ export 'src/sortable/table_sortable_mixin.dart';
 
 export 'src/widgets/flexible_table_content.dart';
 export 'src/widgets/flexible_table_header.dart';
-export 'src/widgets/flexible_table_info_row.dart';
-
-export 'src/widgets/flexible_table_header_cell.dart';
-export 'src/widgets/flexible_table_info_cell.dart';
 
 export 'src/flexible_column.dart';
 export 'src/flexible_table_configurations.dart';